deleted trackController
