+++ server/controllers/authController.js
@@ -28,6 +28,7 @@ LoginUser: async (req,res)=>{
       db = req.app.get('db');
   //check to see if user exists
   const foundUser = await db.check_user({email})
+  console.log(foundUser[0])
   if(!foundUser[0]){
     return res.status(401).send(`User Not Found`)
   }
@@ -42,6 +43,7 @@ LoginUser: async (req,res)=>{
   req.session.user = foundUser[0]
   res.status(202).send(req.session.user)
   console.log(req.session.user)
+
 
 },
     logMeIn: async(req,res)=>{
