+++ src/App.css
@@ -1,19 +1,4 @@
 body {
-<<<<<<< HEAD
- 
-  background-image: linear-gradient(to right, #4e4e4e, #f0f0f0);
-}
-
-.boombox {
-  display: flex;
-  
-  bottom:100vh;
-  opacity: .2;
-  width: 70vw;
-  height: 100vh;
-  float: right;
- 
-=======
   
   background: linear-gradient(to right, #4e4e4e, #f0f0f0);
 }
@@ -22,7 +7,6 @@ body {
   background-image: url('./boombox.png');
   background-size: cover;
   background-repeat: no-repeat;
->>>>>>> master
 }
